fence_compute: Rework how the domain option works
Right now, we always append the domain to the host that we get and hope
for the best. This is not optimal and breaks in the cases where nova is
listing hypervisors with FQDNs but services with short hostnames.

Since the domain option is used when listing the hypervisors to also
list hypervisors with short hostnames, we know that the NovaEvacuate
agent will be able to match both FQDNs and short hostnames.

So what really needs to work fine is when we do actions on the nova
services. We change the logic to actually try with the hostname we're
given (which can be a FQDN or a short hostname); if there's no services
on that hostname, we try to use the domain option to build the
alternative hostname (FQDN if what we had was a short hostname; short
hostname if what we had was FQDN), and work with this instead.

This should allow all mixes of FQDN / short hostnames between nova and
pacemaker.